+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Notes for Shopify
 
+## Unreleased
+
+- Fixed a bug where product titles weren't syncing correctly across sites. ([#198](https://github.com/craftcms/shopify/issues/198), [#201](https://github.com/craftcms/shopify/issues/201))
+
 ## 7.1.2 - 2026-06-03
 
 - Fixed a PHP error that occurred when editing settings in Craft 4. ([#216](https://github.com/craftcms/shopify/issues/216))